Fix where the FontMatrix ends up in faked CFF CIDFonts
If we're making a "fake" CIDFont from a normal CFF font, we use the defined CFF
font as the only entry in the FDArray in the fake CIDFont.

Originally, the fake CIDFont had an identity matrix, and the "real" scaling
matrix was in the actual CIDFont.

This worked fine for the glyph scaling, but caused the metrics to be scaled
incorrectly.

And since one matrix is always identity, there is no risk in simply swapping
them over.