pdfwrite - improve handling of procedural /.notdef in type 1 fonts
Bug #703770 "Postscript to PDF results in Type3 font. Acrobat produces Type1 from same file"

The creating application "Polylogics DIAD White Pages Pagination"
applies a 'ReMapFont' procedure to every font that it uses. This
procedure replaces any existing definition in the font of three glyphs;
/emspace, /enspace and /.notdef, with procedural definitions.

This 'seems to be' intended, for the /.notdef, to make a highly visible
black square on the output, presumably to highlight that there was a
problem.

For pdfwrite this is a problem as we cannot embed a PostScript procedure
in a type 1 font in a PDF file. Ordinarily we would create a type 3
font, execute the procedure and capture the marking operations which we
would then use to create a CharProc. However, the on glyph we absolutely
must be able to embed in an output font is the /.notdef, without that
the font is not valid.

Because we can't copy the /.notdef, we fail to copy the font, leading to
all the glyphs being captured as type 3.

How does Acrobat Distiller resolve this ? It simply ignores the replaced
/.notdef and embeds a CharString which consists of 'endchar'.

We already have code to do something very similar, because the old
ADOBEPS4 PostScript driver for Windows does a similar trick with the
/.notdef.

So this commit extends that 'feature', if we find that the CharString
data for a glyph is not a string, then we test to see if it is the
specific ADOBEPS4 procedure. If it is not then we check to see if the
glyph reference is a name, and if it is whether the name is /.notdef.
If it is the /.notdef then we use the same predefined CharString
instead of the procedure.

I've chosen to leave the ADOBEPS4 test unchanged as I cannot find a test
file to exercise it, and I am not absolutely certain that the glyph
reference will be a name in this case.

This fixes the customer's reduced file, and their original file, in the
sense that pdfwrite now embeds the original type 1 font. If the /.notdef
glyph is actually used (or the emspace or enspace) then these will still
be converted to type 3 glyphs as there is no other way to embed them in
the output. This commit only affects the initial copying of the font.

The Producing application seems to be quite widespread in the production
of directories (white pages and yellow pages I guess) so I'm a little
surprised this hasn't come up before.