Prevent paths being grestore'd
The PDF Reference Manual (section 4.4.1, p226 of the 1.7 edition) states
that the current path is not part of the graphics state and is not saved
and restored along with the graphics state parameters.

This means that any changes to the path between 1 and Q have to persist
whereas in PostScript they are thrown away. To do this we have to copy
the current path before we do a grestore, and assign that copy back to
the current path in the graphics state afterwards.

Fixes /tests_private/pdf/PDFIA1.7_SUBSET/CATX1738.pdf which appears to
be the only file in the test suite which trips over this.