Relax checking of Forms, assume Form XObject if in doubt.
Previously we had handled Form XObjects which were missing a Subtype
by checking for a FormType, and only throwing an error if the FormType
was also missing.

It appears we have several files in the test suite where both keys
are missing (!!) and Acrobat simply assumes these are Form XObjects. So
here we relax the checking and assume that anything missing a Subtype
is in fact a From XObject. Several files now work.