Correctly scale the character spacing (Tc)
For CIDFonts and fonts without a Widths array, we apply the character spacing
value differently - normally, we'll add the Tc value to the value read from the
Widths array. But in the other case, we use the graphics library's feature for
adding a value to the advance width of each glyph in a string.

It turns out, however, that these require different scaling of the width value.