Two fixes for Default* colour spaces
At  the page level, if we fail to create a colour space from the
Default* specification, just carry on leaving the space unassigned
this will use the Device space instead, which is what we're supposed
to do.

When running content streams, save and restore any existing Default*
spaces, and apply any definitions from the stream Resources/ColorSpace
dictionaries.

We can't, unfortunately, simply do this in one place because Page
Contents can be arrays not single streams, so we need to ensure
we have the Default spaces in place before we start the stream, whereas
Forms and Patterns can be handled at the stream level in one place
rather than writing almost exactly the same code in 2 places.