Fix colorspace of forms with transparency.
Need to save the colorspace before starting the form group, in case it
gets changed by the transparency code.

Add parameter to pdfi_form_execgroup() to override the colorspace if needed.

We have different paths through this code, that need to handle the
graphics and color space different.  This seems to be the right way to
do it (and might possibly match what the gs code does, if the gs code
wasn't so unreadable that who can tell for sure?).