Fix some stack/reference problems
When reading xrefs we leave a dictionary on the stack and use it from
there rather than taking a reference to it.

In a few places we weren't popping it from the stack, and in a few
others we were popping too many objects leading to a stack underflow.

This was causing errors with /tests_private/comparefiles/Bug690170.pdf