Attempt to fix indeterminism in output of pcx16 device.
The existing code overreads the end of the line. That's probably
OK in terms of not causing crashes (as we only overread by enough
to make it up to the next 32bit word), but is enough to read
undefined input and so a) upsets valgrind, and b) gives indeterminate
output.

Fix it by handling the last few pixels with extra checks.