Fix AES encrypted streams with incorrect Length
We need to use a SubFileDecode filter on encrypted streams, because we
can't use endstream/endobj to terminate the filter. This is because
the decryption will corrupt the strings.

We can't use EODString, because that seems to limit the maximum length
of returned data to the length of the EODString, which doesn't work with
AES decryption because it needs a minimum of 16 bytes at a time.

So we must have the Length and it must be correct. The file
/tests_private/comparefiles/Bug692852.pdf is AES encrytped and has a
stream with an incorrect Length. Because the Length was shorted than
the actual stream we were not reading all the data from it leading to
missing content.

So, if we find a stream whose Length is invalid, this commit changes
our behaviour to laboriously trawl through the file looking for either
an 'endstream' or 'endobj' keyword, and calculate the correct Length
from that, which we then store in the stream (and mark it as valid).

This is slow, but it will only happen for invalid files.

Fixes /tests_private/comparefiles/Bug692852.pdf