Solve indeterminism in pr1000_4 device.
There are 4 devices (pr201, pr1000, pr150, and pr1000_4) that are
all basically the same code. They differ in the number of lines
of data they deal with at once.

The first 3 use 24, 40 and 48 lines respectively - all numbers
that are multiples of 8. The code reads that many lines at once
and then "transposes" the data to send 'columns' of 24/40/48
pixels at once.

The pr1000_4 variant uses 60 lines. This means we read outside
the input buffer into uninitialised data.

I have my doubts that this device actually works properly, but
the change here simply bumps up the buffer sizes to be larger
and initialises it all to 0 so we get consistent SHAs in cluster
testing.