Bug 703788: Bounds check name table "string"
When parsing a Truetype font we store the name table data in a Postscript string
and later try to extract a font name we can use for the Postscript FontName.

We were checking for a zero length table, but not checking whether a non-empty
string was long enough to contain the requested name index.

So handle that situation gracefully.