Fix Bug 703774: Overprint+Transparency+Fill_Stroke problems.
The customer file showed a problem stemming from Fill_Stroke, causing
missing text in clist mode was caused by the pdf14_mark_fill logic using
the stroke_overprint for the fill, with drawn_comps not properly set.

The underlying issue is the pdf14_device->op_state not being correct in
some cases. Fixing just the fill_stroke case showed up differences in other
files due to similar confusion/mis-match between the color, the drawn_comps,
and the overprint setting, evident with Bug693220.pdf.

This fixes all of that and also made some 'hygiene' changes to avoid assigning
the overprint and pdf14 device op_state from a boolean (using the correct
enum).

Lastly, fix the setupOPtrans in the PDF interpreter so that the stroke
overprint is taken into account if the operation is one of the fillstroke
operations (previously it only used the filloverprint setting).