pdfwrite - defer freeing annotations which are named objects
Bug 703783 "Segmentation fault with named ANN pdfmark"

Ordinarily we write annotations immediately they are created, free the
contents immediately and free the dictionary when we write the Annots
array in the page dictionary.

However, if the Annot is also a named object, then it is additionally
referenced from the local named resources dictionary. In this case we
cannot free the annotation dictionary when we write the Annots array.
Because it is a named object it could still be altered.

Instead we must defer freeing both the annotation contents and the
Annot dictionary until we free the local named resources (when we close
the output file).