Don't error on broken escaped names
Names which contain white space must and characters outside the range
33-126 should, use the name escape sequence #xx to enter the code as a
2-digit value.

Obviously the name escape character '#' must also be escaped itself in
the same way if it is used directly.

Test file /tests_private/comparefiles/Bug693570.pdf has an annotation
with a name (/#custom) which is not escaped and so is illegal. However
Acrobat (as usual) can cope with it.

Deal with it here by ignoring errors in escaping names, but flag a
warning.