Ensure dereference of otional content Usage data
Customer file, the file has two Optional Content Groups with /Usage
entries, and the /Print dictionary is an indirect reference. It looks
like we've run across this before in different guises; the value
associated with the View key is forcibly dereferenced, but not the value
associated with the Print key, whereas the value of the PrintState
key is forcibly dereferenced, but not the value associated with the
ViewState.

This commit rationalises all of these by focribly dereferencing all the
various objects.

This solves the problem for the customer file and potentially solves
a problem if the ViewState should be an indirect reference.