Much work on halftones
Add code to create and populate a gx_transfer_map from a pdfi
transfer function object.

Use the pdfi_evaluate_transfer() function to create transfer maps for
type 1, 10 and 16 halftones as required (if /TransferFunction is
present in the halftone dictionary and is a stream object).

Fix an error in type 10 halftones that was leaving two values
uninitialised and could cause a seg fault.

Fix an error requiring Width2 and Height2 entries in type 10 halftone
dictionaries (these are optional) which could lead to an error and the
halftone not being applied.

There may be more work required in this area, but this works for all
the test cases I've been able to find or create.