Fix clist phase problem.
If we are strip_copy_rop2ing a tile, ensure that we set the
tile phase, whether or not we find it in the cache.

This causes various diffs in the cluster, but manual testing
shows that in all the ones I tested (more than half) they now match
the page mode rendering.