Add some preprocessing of AcroForm in trailer funcs
Put AcroForm and NeedAppearances in context for efficiency

This code gets called once per doc, as compared to the AcroForm and
Annotation code, which gets called later, once per page.

There is a TODO: regarding creating the ParentField (which is a hack I
don't want to do) and Bug692447.pdf that would apparently be handled
here, but I haven't looked at it yet.