PDF interpreter - set defaults for FreeText annotations
Set up some minimal appearance defaults for FreeText annotations, in
case the Default Appearance (/DA) doesn't do so. This is to fix a
customer file where the /DA set the colour to red, but did not select
a font or point size, leading to no text being drawn.

Previously we assumed the /DA would set all of the required graphics
state and only set up these defaults if the /DA was not present.