Add CIDFont fall back substitution
(Currently all non-embedded CIDFonts use the fallback.)

This also adds support for begin/endbfchar and begin/endbfrange cmap operators
often used by ToUnicode CMaps, and support for ToUnicode CMaps.

Also adds decoding tables for CNS1, GB1, Japan1, Japan2, and Korea1 orderings,
to convert CID values to Unicode values.

So, when a substitute CIDFont is in use, we'll use the ToUnicode CMap to lookup
a Unicode code, if there is no ToUnicode or it doesn't contain a code, if the
CIDFont ordering is one of the above 5, we'll use the decoding table to find
a Unicode code, then use the TTF font 3.1 cmap to map Unicode to GID.

If there is no ToUnicode, and not available decoding, we'll just believe the
CMap mapping, and treat that as the GID.

TODO: make the pdfi "PS" font parser stack extendable - ToUnicode CMaps, it
turns out, play fast and loose with the alleged limited stack requirements to
parse CMaps