Set decrypt_strings true when dereferencing normal objects
If we have previously set decrypt_strings to false (because we are
processing a stream) then we need to set it back to true when
dereferencing any objects accessed by the stream operators.

We already were setting it false when the referenced object was in a
compressed ObjStm, and preserving the state, so this is a simple change.