Move much of the memory cleanup currently done in pdfi_free_context into a new routine, pdfi_clear_context, which cleans up all the entries in the context which depend on the current PDF file (eg Root, Pages etc). This should allow us to start a new file after closing the old one.
Call the new pdfi_clear_context() routine from both pdfi_free_context
and from pdfi_close_pdf_file()

Modify a couple of routines which free stuff to set the stored pointer
in the context to NULL afterwards so we don't attempt to double free
them.

Add a 'high water mark' depth of gsaves, initialised in
pdf_impl_set_device(). When cleaning up in pdfi_clear_context gsave
back to this point, not all the way back to the initial gsave. This
prevents us grestoring back to a point before the device was installed.

Add code to read the Collection dictionary from the Root dictionary,
if present.

Move the code to process each page in a PDF file to a new routine
outside of pdfi_process_pdf_file(). If we have no collection code then
call the new routine, pdf_process() to process all the pages.

If we do have a collection then call a new routine
pdfi_process_collection() to extract each embedded file, check to see if
it is a PDF file, if it is write it to a temporary file. Then call
pdfi_process_pdf_file to handle each embedded PDF file. This may be
recursive if any embedded file is also a collection. If there are no
embedded PDF files, then just call pdfi_process() to proceed with the
containing file as if it were not a collection.

Future enhancement : It should be possible to do all this juggling
without actually writing temporary files. However we would need a
way to save the current file 'state'; that is all the things which
we are clearing down in pdfi_clear_context. We would need a stack
of 'states' because a PDF file embedded in a collection could
itself be a PDF collection.