Bug 703829: Use an array for the outline in a PDF Ink annotation
Previously, the Postscript would set a mark, push all the coordinates onto the
op stack, then call the custom operator (.pdfinkpath) to create the path.

In the fuzzing file, the number of coordinates is sufficient for the op stack
to require extending, meaning we can't simply index "down" the stack to
retrieve all the coordinates - we can only safely, directly access the elements
of the current "extension".

To make this more reliable and robust, store all the coordinates into an array
(in Postscript), and have .pdfinkpath access them from the array.