Make the "PS" stack dynamic
The stack for the minimal Postscript-like interpreter in pdfi was originally
sized to handle interpreting conventional CMaps, which should never have more
than about 320 objects on the stack - the spec is specifically written to limit
the maximum stack usage. That also allowed more than enough stack space for
(valid) Type 1 font creation.

Turns out, however, that ToUnicode CMaps play fast and loose with that aspect of
the CMap spec, and occasionally push a lot more objects.

So, make the stack expand if required. The initial size is still plenty for well
behaved CMaps and Type 1 fonts, but can expand to cope with badly behaved
ToUnicode CMaps.