Move cs_adjust_color_count (and twin) from being macros to functions.
This also involves them moving into gxgstate.h due to the need for
the gx_currentcolorspace_inline (etc) macros to be defined before
the function can be. This is merely a more explicit form of the
dependency that already existed, because the macros could not be
used without that header being included in the current code.