Annotation enhancements -- AP and field /CH
If there is no /N AP, then find /R or /D.  This is what gs does.
Also, apparently not finding the matching AS doesn't seem to be an
error, it just means don't use an AP?

Implement field /Ch

This is a text choice field, and gs implements it basically the same
as Tx, except making sure to regenerate the AP if NeedAppearances is
set.