Fix breakage due to getbits -> get_bits_rectangle change.
When we used both get_bits and get_bits_rectangle, we had code
in there to protect us against infinite recursions. When we moved
to just using get_bits_rectangle, I removed that, on the mistaken
assumption that we'd no longer need it.

It turns out that we can still hit cases where it's important.

For example, with:

 bin/gs -Ilib -sOutputFile=out.pxl -dUseCIEColor -sDEVICE=pxlcolor
        -r300 -Z: -sDEFAULTPAPERSIZE=letter -dNOPAUSE -dBATCH -dClusterJob
	-dJOBSERVER /home/regression/cluster/perf-testing-gpdl/ps/Ad_Quark.ps

On page 3, we have an image that is plotted through the simulate
overprint stuff. That tries to 'get_bits_rectangle' to read the
existing pixels, but because pxlcolor is a high-level device
it goes to gx_default_get_bits_rectangle, and goes into an infinite
loop. This only happens when UseCIEColor is in force.

I have reenabled the infinite recursion protection here.

Possibly pxlcolor should be defining its get_bits_rectangle entrypoint
as being gx_no_get_bits_rectangle?

Now both UseCIEColor and non UseCIEColor runs die with the same
'undefined in imimage' message.