Have psdcmykog initialize its colorants to 6
The psdcmykog device should have its max_separations and its number of
components be the same.  This means it will not add extra components
when they are encountered in a document like psdcmyk and tiffsep
do, which are also devn device.  psdcmykog device is unique in that
it is a devn device but with a fixed set of colorants.