Rename opmode to opmsupported, and tweak operation.
First off, we rename 'opmode' to 'opmsupported'. opmode does NOT
mean the same as OPM in PDF, but rather indicates whether OPM
can be supported on a given device or not. Renaming should solve
this confusion.

On creation of a device dev->color_info.opmsupported is set to
GX_CINFO_OPMSUPPORTED_UNKNOWN. The idea is that this UNKNOWN value
should be resolved to either 0 or 1 (OPMSUPPORTED_NOT or OPMSUPPORTED)
before it is used.

This requires code to check for it being unknown, and if it is,
call a function to resolve it.

We have found at least one place in the code where this fails to
happen properly. To protect against this happening elsewhere,
we introduce a gx_get_opmode function that performs this
"Is it unknown? If so resolve. Read value." dance for us. By insisting
that we always call this rather than accessing directly, we harden
ourselves against missing cases.

This produces a far few diffs, all of which look like progressions to me.