Support Multiple Master Type 1 fonts
Add just enough support for MM fonts so Freetype believes it is a MM font, and
handles it accordingly.

Mostly, what's left out is related to hinting, which we disable anyway.

Possibly the remaining thing to support would be the Blend/ForceBold key.

Also fix memory leaks with XUID values.