Revert "Small tidyups within bbox device." and improve comments.
This reverts commit 5c433fa5edbb10efff7518f69779a319c8ea342b.

It turns out that the old (now restored!) code within gdevbbox.c
handles the dev->target == NULL case differently to the forwarding
device.

Update the comments in gdevbbox.c that mislead me into making this
change in the first place.

Thanks to Ray for spotting the issue, and Ken/Chris for talking it
through with me.