Fix XUID values leak
The code to finalize the font cache, and to remove font/matrix pairs from the
cache had never been updated to cope with XUID dynamically allocated values,
it only freed the memory pool used for the font/matrix pairs, ignoring the
xvalues for a XUID, which is not allocated from that pool.

Spotted working on pdfi, which is the only interpreter where XUIDs are likely
to arise, and that doesn't rely on garbage collection.