Process PDF collections from the PostScript interpreter
We added support for PDF collections in a previous commit, but only for
the native pdfi interpreter. This commit adds the ability to handle
these when called from the PostScript environment.

We still (for now at least) write the embedded files to disk as scratch
files, and then process those as individual PDF files. Ideally we'd like
to do away with creating the scratch files, but that is a future
enhancement, which will require changes to the PostScript processing
as well.

The /Collection entry (if present) in the dictionary returned by
.PDFInfo contains an array of strings, twice the size of the number of
detected PDF files. The first string is the temporary filename, the
second is the name tree entry for that file. Currently we don't
use that, but PostScript programs could use it to select a specific
file for processing.