Fix patterns with type 3 coloured fonts
Patterns inherit their graphics state, and in particular the CTM, from
the 'enclosing context', when in fact they need to use the CTM of the
type 3 font (which includes the font scaling).

By using the pdfi_run_context() function instead of
pdfi_interpret_inner_content_stresam() directly we get that benefit and
should be protected against similar functionality updates if we find
anything else wrong like this.

Fixes 2 test suite files, these appear to be the only files we have
which exercise this.