pdfwrite - Do not lie about emitting Multiple Master fonts
The pdfwrite and ps2write devices do not write Multiple Master fonts;
instead we write a type 1 equivalent of the instance of the MM font.

However, we were setting the Subtype of the font to MMType1, which is
obviously wrong (though it doesn't seem to cause any problems).

Fix that here by always setting the Subtype to Type1 (for simple fonts
which are not TrueType).