Disable overprint for images (like gs does)
This has been a bug since forever, but it mostly matters in separation
colorspaces such as such as psdcmyk and pkmraw, which have not been
heavily tested.

It was noticed in this file for pgmraw, when Robin's recent overprint
changes apparently changed some behavior in this file:

tests_private/pdf/uploads/PDFX-ready_Output-Test_30d_X4.pdf

These were (probably) the recent changes, just for posterity:
https://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=f998654cb67f142791db43629d3c9f4600380701
and
https://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=5141ae221771ed0e95b3a587a149b68c834f2bf2

Quite a few progressions, mostly in pkmraw, psdcmyk files that do things with
overprint and images.