Fixes images with both SMask and Mask entries
If we had an image with both an SMask and a Mask key, then we would
not retrieve the Mask params, which led to the Mask image throwing an
error, and the image not being rendered.

There doesn't seem to be any reason why we should not fill in the
mask_info in the presence of an SMask (we do try to execute the Mask
image anyway), and it makes the file
/tests_private/pdf/sumatra/1901_-_tiling_inconsistencies.pdf render
correctly without causing any regressions.