Fix retrieving the page size from the device.
When we call pdfi from gs, we don't set the page size directly from the PDF, we
let Ghostscript do that, and then retrieve it from the device so pdfi knows
the right dimensions.

The code to retrive it was undoing scaling that it didn't need to, causing
problems, especially with transparency and clist.