More repair improvements
pdfi_repair_file now does not close or reopen the main stream, which
means we can now call it from places that can't have the main stream
closed (eg in a content stream).

It also now returns the main stream to the point it was at before the
attempted repair.

When dereferencing an object, if we fail to read a valid object call
pdfi_repair_file(), and then try again. If the error was caused by an
incorrect offset this should fix the problem.

Note that we only ever repair files once!