Fix Page size for transparency when pdfi is running inside GS
We still use the page size for transparency bounding calculations,
even when the media is set by Ghostscript not pdfi. However the
transparency clipping seems to be applied in user space, not device
space, so we need to take any translation into account.

This solves a number of problems with files being clipped.