More number parsing fixes to more closely match Acrobat
Seems Acrobat is happy with any 'number' which is composed of digits,
decimal point, e and + or -, in any order and any frequency.

malformed numbers such as 80.abc cause Acrobat to display an error.

In addition to the previous commit this commit treats numbers of the
form:

40.-40

As if the characters after the '-' were not present. This 'seems to be'
what Acrobat does.

The file /tests_private/pdf/sumatra/2231_-_invalid_characters_after_number.pdf
now renders the same as Acrobat.