XPS interpreter - Undo CTM scaling when preserving Tr 2
The scaling of the stroke width when rendering text in mode 2 (fill and
then stroke) works for rendering but when preserving the text rendering
mode for high level devices it ends up scaling the width too much.

By applying the horizontal scaling from the CTM, when preserving the text
rendering mode,  we return the width to a reasonable value without
affecting rendering.