More media size changes; fix NoUserUnit and PDFFitPage
Fix NoUserUnit; even if we aren't applying UserUnit, we still need to
deal with the Box size.

If we are fitting the PDF to the existing media, we no longer need
to return the media size. I don't think this works properly currently
(at least with rotated pages) but it no longer fails with an error now.