Fix device setup with NumSpots and Transparency
These are required in order to correctly allocate memory; in particular
the clist fails to erase pages properly if these are incorrect.

While this works for the simple case of rendering one page from the
PDF file at a time, if we ever want to run multiple PDF pages to one
PostScript canvas, we will need to retrieve this information from all
the PDF pages, consolidate it, and set it up in advance of running any
page. This will require checking for spot name duplicates which we
cannot currently do, because we don't yet return the ink names.