Yet more work on transsarency clipping called from PostScript
As (I think) noted before, we use the page.Size to set up bounding
boxes for the transparency compositor, which it clips to, and these are
defined in user space, not device space. We still weren't getting that
correct.

In this commit we basically duplicate much of what the non-PostScript
case does; which seems to work for the few cases that were previously
wrong. I'm not entirely certain why this doesn't fail in the native
ghostpdf interpreter, but it seems OK there. Possibly we will find out
later that there's a problem and have to fix that too.