Fix stroke width on stroking text rendering modes
When stroking text (due to a text rendering mode) we must adjust the
linewdith, because we alter the CTM by multiplying in the text line
matrix and the font size.

But due to misplacing the assignment, we were also adjusting by the
existing CTM....

At 72 dpi this has no effect, but as the resolution increases the stroke
width gets smaller, leading to incorrect rendering.

Calculate the modified stroke width before we multiply in the CTM.