Yet another attempt to get the page size correct for transparency
The translation being applied to the page needs to be accounted for too,
but unlike the media size, it is affected by the CTM, so if the media
is rotated we need to apply it differently.

I've got about 13 files which are affected by this problem and all
render correctly with this fix, but I'm not confident about this.