Setup strokeadjust to match gs
'strokeadjust' needs to be false for 'high res' devices and low for
'low res' devices.  This causes subtle difference in the way some
lines are drawn.

Also support the DITHERPPI argument.

Note that this code is in the device setup for pdfi.  It won't help the
pdfi-in-gs case.  That case also has the wrong values, because it is supposed
to be inheriting the graphics state from the gs interpreter, but it currently
does not.

gs_gstate_alloc() will set strokeadjust to 'true', so that will be the
value if it's not overridden.

Sample: I trimmed down page 65 of
tests/pdf/Bug6901014_Arioli-NAG-Warwick.pdf to be just the black
horizontal line at the top of the box.  This affects many samples.