Expand and alter support for NCLR ICC profile and color names
This commit affects the capabilities of the psdcmyk device
and the tiffsep device.  Much of the discussion below
is now in the documentation, but is included here for completeness.

Normally with an NCLR ICC profile you can specify the
colorants using
-sICCOutputColors="Cyan, Magenta, Yellow, Black, Orange, Violet"
for example for a 6CLR ICC profile.

Now we allow additional names beyond those of the ICC profile
component count.  In this case, those names will be installed
into the tiffsep or psdcmyk device list of colors, following
the ICC profile colors.

The number of spot colors (those that go beyond the standard CMYK colors)
allowed by tiffsep or psdcmyk can be set using -dMaxSpots=#.  The default
value for this is currently set to 10 (GS_SOFT_MAX_SPOTS).

As an example consider the case where we wish to use a
6CLR ICC profile that includes Orange and Violet, but
need the device to include a specialty color such as
Varnish, which does not appear in the document and is not
handled by the 6CLR ICC profile.  In addition, we desire
to allow one more spot color of the document to come through
to our device.  For this case using
 -sICCOutputColors="Cyan, Magenta, Yellow, Black, Orange, Violet, Varnish"
   -dMaxSpots= 4
   -sOutputICCProfile=My_6CLR_Profile.icc
would provide the desired outcome.

Note that it is up to the device or through the use of
-sNamedProfile to involve the setting of any values
in the Varnish channel.  However, if an All color value
is encountered in the document, the Varnish component will
have its value set as will the Orange and Violet components
(Likewise if a spot color named Varnish is
encountered in the document the Varnish component will
be used for the values).

Finally this commit adds a debug option --debug=clist-color
which will provide clist writing and reading debug information
for the reading and writing of devn color values to and from
the clist.  This is helpful to debug cases where the pdf14 color
model changes due to group color space changes, which can be
difficult to track down when there is a mismatch in the read and
write of the color values.

Thanks to Ray Johnston for his help on much of this effort.