When running from PostScript, use the PS gstate
We are supposed to use the PostScript graphics state when run from the
PostScript environment, but the act of creating a PDF interpreter
context created a new graphics state, which we used.

This commit has the create context function take a pointer to an
initial gstate, which it uses if present. Otherwise we create a new
graphics state, as before.

Also update the clear and free context functions so that they don't
attempt to throw away anything being used by the initial graphics
state.

Modify the PDF interpreter to call create_context with a NULL pointer
and the PostScript interface code to call it with the PostScript
graphics state.