Cope with Type 0 fonts reusing descendant fonts.
Originally, the code didn't account for the possiblity of a
composite font that shared a descendant with another composite
font, so didn't cope with the object lookup return a PDF_FONT
object rather than a PDF_DICT object.

So rejig things so we can cope. This also meant tidying up
the CIDFont definitions into a common "pdf_cidfont_t" type
as well as the two type sepcific (CIDType0 and CIDType2) ones.

It also meant moving the flag that identifies a descendant as
a subsitute font from the Type 0 font object into the CIDFont
object.