Vector devices - fix compiler warning
Bug #703866 "./devices/vector/gdevpdfd.c:447: warning: ‘gdev_vector_dopath_segment’ accessing 24 bytes in a region of size 8 [-Wstringop-overflow=]"

Benign warning. For the cases where we are passing a pointer to a single
fixed point number the type of operation only consumes a single
number. I'm not at all sure (and this code is > 20 years old) why the
original author chose to define it as a pointer to an array of 3.

This is very minor. However since we aim to keep compiler warnings to
a minimum it is addressed in this commit by changing the function
prototype.