Fix bug 703848: Missing interior of rectangles (regression in 9.54.0)
This was caused by the commit bdf7cf8b6c. If overprint was true, in the
sections of the 8 and 16 bit do_mark_fill_rectangle_ko_simple that copy
the 'dst' pixel to the plane buffers (dst_ptr), the colors were copied,
but not the alpha.

For the example file, the fill_stroke had the fill end up transparent
and the stroke be visible, since stroke_overprint was false.