Param parsing - Fix string lifetime problem
Bug #703895 "pages duplication error"

The command line reuses exactly the same parameter to PageList, this is
a problem because we end up freeing the string from the parameter list
before we use it.

How does this happen ? Well, its complicated. The first time we get the
string sent we create memory for it, and store it. Then we get a 'get'
params, for which we manufacture a param string, using the stored data.

When the same string is then set, the PostScript interpreter uses the
string returned from the 'get' params (because its unchanged), and the
first thing we do, given a 'new' string, is free the old string data.
But the data portion of the new string is pointing at the old string
data, resulting in it being freed while we are still trying to use it.

Generally this isn't a problem (through sheer luck), but should the free
trigger a garbage collection, its possible for the old string data to
be overwritten, resulting in 'nonsense' which we then try to parse.

The easy solution is to create a transient string for return to the
PostScript environment, which means that the PostScript world won't
retain a pointer to it, and will create a fresh string for the second
'put' params.