pdfwrite - limit co-ordinates for PDF/A output
Bug #703896 "Regression: some PDF/A conversions from PostScript have invalid float range"

Commit 3156b0aa8ac4e9ef440c169799658da249eb41c9 removed a long-standing
'kludge' which limited the co-ordinate values we could use to the range
permitted by Acrobat 4; +/- 16384.

Given that the piece of software in question had been obselete for 15
years, this seemed reasonable, but it turns out that, buried in the
PDF/A-1 spec is a small note; 6.1.12 'Implementation limits' which
specifies (rather horrifyingly) that a conforming file must not violate
the limitations of Acrobat 5. Not the PDF 1.4 specifications, but the
actual limits of that particular piece of software.....

So, here we reinstate the code to work around the limitations; but
we use the limits of Acrobat 5 (+/- 32767) rather than Acrobat 4. We
also use the production of PDF/A-1 (specifically version 1, version 2 is
not subject to these limits) rather than the production of PDF 1.4 to
trigger the implementation.

In the places where we clamp a value rather than scaling it, we consult
the PDFACompatibilityLevel to see what to do.

As the original commit noted; I'm far from certain that this covers all
the places where we might possibly emit co-ordinates outside the
range of the architectural limitations of Acrobat 5, but it does restore
the functionality so that it as least no worse than it was before the
commit to remove it.