Handle another case of broken Postscript
This time in a ToUnicode CMap.

This CMap uses "<<" to start a dictionary (not supposed to be supported), fails
to close the dictionary again (no ">>") and fails to use a leading "/" to
identify a name (a name we ignore anyway),

Since we don't really bother with sub-dictionaries, just ignore "<<" and
(for completeness) ">>".